day7_weather_bot/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.135.3
fonttools==4.62.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.4
openai==2.31.0
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
